--- sdk/watchtower-sdk/src/types/index.js
@@ -1,3 +1,0 @@
-export * from "./error_interface";
-export * from "./log_interface";
-export * from "./performance_interface";
